// frontend/src/api/model/getFilesOrderBy.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GetFilesOrderBy = typeof GetFilesOrderBy[keyof typeof GetFilesOrderBy];


export const GetFilesOrderBy = {
  created_at: 'created_at',
  size: 'size',
  name: 'name',
} as const;